requirements.txt
requests
beautifulsoup4
python-dotenv
supabase
httpx
pdfplumber
google-genai
litellm
pydantic
playwright
markdownify
reportlab
playwright-stealth